like: enhanced to replace `matches`
1. enhanced like to support multiple modes:
   a. String match
   b. Lambda match
   c. Object match
2. remove `matches` condition
