+++ app/core/src/main/scala/com/qwery/runtime/instructions/queryables/Expose.scala
@@ -8,7 +8,7 @@ import com.qwery.runtime.datatypes.{BooleanType, StringType, TableType}
 import com.qwery.runtime.devices.RecordCollectionZoo.MapToRow
 import com.qwery.runtime.devices.RowCollectionZoo.createQueryResultTable
 import com.qwery.runtime.devices.{QMap, RowCollection, TableColumn}
-import com.qwery.runtime.instructions.conditions.Matches
+import com.qwery.runtime.instructions.conditions.Like
 import com.qwery.runtime.instructions.expressions.TableExpression
 import com.qwery.runtime.instructions.functions.{AnonymousFunction, FunctionCallParserE1, ScalarFunctionCall}
 import com.qwery.runtime.instructions.queryables.Expose.exposeMatch
@@ -22,14 +22,14 @@ import qwery.io.IOCost
  * @param expression the [[Expression expression]] to expose
  * @example {{{
  *   set response = { id: 5678, symbol: "DOG", exchange: "NYSE", "lastSale": 90.67 }
- *   response matches { id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber }
+ *   response like { id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber }
  * }}}
  */
 case class Expose(expression: Expression) extends ScalarFunctionCall with RuntimeQueryable with TableExpression {
 
   override def search()(implicit scope: Scope): (Scope, IOCost, RowCollection) = {
     expression match {
-      case Matches(source, pattern) =>
+      case Like(source, pattern) =>
         val rows = exposeMatch(src = QweryVM.execute(scope, source)._3, pattern = QweryVM.execute(scope, pattern)._3)
         val out = createQueryResultTable(returnType.columns)
         val cost = out.insert(rows.map(_.toRow(out)))
@@ -56,7 +56,7 @@ object Expose extends FunctionCallParserE1(
        |isUUID = v => v.isUUID()
        |isNumber = v => v.isNumber()
        |response = { id: "a891ee9b-6667-40fc-9ed1-a129d04c8b6d", symbol: "ABC", exchange: "NYSE", lastSale: "35.76" }
-       |expose(response matches { id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber })
+       |expose(response like { id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber })
        |""".stripMargin) {
 
   def exposeMatch(src: Any, pattern: Any)(implicit scope: Scope): List[QMap[String, Any]] = {
