+++ app/core/src/test/scala/com/qwery/runtime/instructions/conditions/LikeTest.scala
@@ -1,7 +1,7 @@
 package com.qwery.runtime.instructions.conditions
 
 import com.qwery.language.models.Expression.implicits._
-import com.qwery.runtime.Scope
+import com.qwery.runtime.{QweryVM, Scope}
 import org.scalatest.funspec.AnyFunSpec
 
 class LikeTest extends AnyFunSpec {
@@ -27,6 +27,62 @@ class LikeTest extends AnyFunSpec {
 
     it("""should decompile: like(name, "Lawr%")""") {
       assert(Like("name".f, "Lawr%").toSQL == """name like "Lawr%"""")
+    }
+
+    it("should evaluate: 'Hello World' like 'H% W%'") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|"Hello World" like "H% W%"
+           |""".stripMargin)
+      assert(result == true)
+    }
+
+    it("should evaluate: 5678 like isNumeric") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|isNumeric = (o: object) => o.isNumber()
+           |5678 like isNumeric
+           |""".stripMargin)
+      assert(result == true)
+    }
+
+    it("should evaluate JSON objects (match)") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|val response = { id: 5678, symbol: "DOG", exchange: "NYSE", lastSale: 90.67 }
+           |val isExchange = (s: String) => s in ['NYSE', 'AMEX', 'NASDAQ', 'OTCBB']
+           |val isNumber = (o: Any) => o.isNumber()
+           |val isString = (o: Any) => o.isString()
+           |response like { id: isNumber, symbol: isString, exchange: isExchange, lastSale: isNumber }
+           |""".stripMargin)
+      assert(result == true)
+    }
+
+    it("should evaluate JSON objects (no match)") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|val response = { id: "5678", symbol: "DOG", exchange: "NYSE", "lastSale": 90.67 }
+           |val isNumber = (o: Any) => o.isNumber()
+           |val isString = (o: Any) => o.isString()
+           |response like { id: isNumber, symbol: "DOG", exchange: isString, lastSale: isNumber }
+           |""".stripMargin)
+      assert(result == false)
+    }
+
+    it("should evaluate an array of JSON objects (match)") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|val response = [{ id: 5678, symbol: "DOG", exchange: "NYSE", "lastSale": 90.67 }]
+           |val isNumber = (o: Any) => o.isNumber()
+           |val isString = (o: Any) => o.isString()
+           |response like [{ id: isNumber, symbol: isString, exchange: isString, lastSale: isNumber }]
+           |""".stripMargin)
+      assert(result == true)
+    }
+
+    it("should evaluate an array of JSON objects (no match)") {
+      val (_, _, result) = QweryVM.executeSQL(Scope(),
+        """|val response = [{ id: "5678", symbol: "DOG", exchange: "NYSE", "lastSale": 90.67 }]
+           |val isNumber = (o: Any) => o.isNumber()
+           |val isString = (o: Any) => o.isString()
+           |response like [{ id: isNumber, symbol: isString, exchange: isString, lastSale: isNumber }]
+           |""".stripMargin)
+      assert(result == false)
     }
 
   }
