+++ app/core/src/test/scala/com/qwery/runtime/instructions/queryables/ExposeTest.scala
@@ -14,7 +14,7 @@ class ExposeTest extends AnyFunSpec {
            |val isString = (x: Any) => x.isString()
            |
            |val response = [{ id: '123456789', symbol: 'AAPL', exchange: 'NYSE', lastSale: 87.99 }]
-           |expose(response matches [{ id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber }])
+           |expose(response like [{ id: isUUID, symbol: isString, exchange: isString, lastSale: isNumber }])
            |""".stripMargin)
       assert(results.toMapGraph == List(
         Map("expression" -> "(x: Any) => x.isUUID()", "value" -> "\"123456789\"", "result" -> false),
